Update the licences accordingly.
Technically the "creator" should be the first author.
I'm just a maintainer of this modified fork.

However, in order to avoid confusion and have people send bug reports
about this fork to the wrong persons, I'm putting myself as the
"creator" and the previous "creator" as a "contributor".

If that generates too much drama, I'll revert this change.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>